requirements.txt
openai>=1.0.0
sounddevice
numpy
python-dotenv
websockets
asyncio
pytest